+++ src/routes/computePlans/ComputePlans.tsx
@@ -195,6 +195,24 @@ const ComputePlans = (): JSX.Element => {
         setDownloading(false);
     };
 
+    const onRefresh = () => {
+        dispatchWithAutoAbort(
+            listComputePlans({
+                page,
+                match,
+                ordering,
+                status,
+                key,
+                creation_date_after: creationDateAfter,
+                creation_date_before: creationDateBefore,
+                start_date_after: startDateAfter,
+                start_date_before: startDateBefore,
+                end_date_after: endDateAfter,
+                end_date_before: endDateBefore,
+            })
+        );
+    };
+
     return (
         <TableFiltersContext.Provider value={context}>
             <VStack
@@ -219,9 +237,18 @@ const ComputePlans = (): JSX.Element => {
                             <StartDateTableFilter />
                             <EndDateTableFilter />
                         </TableFilters>
-                        <SearchBar />
+                        <SearchBar placeholder="Search name or key..." />
                     </HStack>
                     <HStack spacing="2.5">
+                        <Button
+                            size="sm"
+                            variant="outline"
+                            onClick={onRefresh}
+                            isLoading={computePlansLoading}
+                            loadingText="Loading"
+                        >
+                            Refresh
+                        </Button>
                         {HYPERPARAMETERS && (
                             <CustomColumnsModal
                                 computePlans={computePlans}